ziprecruiter/requirements.txt
requests
beautifulsoup4
flask
pika
gunicorn
redis
python-dateutil
pytz